requirements.txt
streamlit
pandas
plotly
mysql-connector-python
python-dotenv
groq